+++ js/config_constants.ts
@@ -117,10 +117,6 @@ export type OnfidoFaceVideoCapture = {
 export type OnfidoFaceMotionCapture = {
     type: OnfidoCaptureType.MOTION;
     recordAudio?: boolean;
-/**
- *  @deprecated Motion is now supported on all devices. There is no need to provide a capture fallback as it will not be applied.
- */
-    motionCaptureFallback?: OnfidoFaceSelfieCapture | OnfidoFaceVideoCapture;
 };
 
 export enum OnfidoCountryCode {